+++ examples/classification/plot_lda_qda.py
@@ -13,8 +13,8 @@ import matplotlib.pyplot as plt
 import matplotlib as mpl
 from matplotlib import colors
 
-from sklearn.lda import LDA
-from sklearn.qda import QDA
+from sklearn.discriminant_analysis import LinearDiscriminantAnalysis
+from sklearn.discriminant_analysis import QuadraticDiscriminantAnalysis
 
 ###############################################################################
 # colormap
@@ -125,18 +125,18 @@ def plot_qda_cov(qda, splot):
 
 ###############################################################################
 for i, (X, y) in enumerate([dataset_fixed_cov(), dataset_cov()]):
-    # LDA
-    lda = LDA(solver="svd", store_covariance=True)
+    # Linear Discriminant Analysis
+    lda = LinearDiscriminantAnalysis(solver="svd", store_covariance=True)
     y_pred = lda.fit(X, y).predict(X)
     splot = plot_data(lda, X, y, y_pred, fig_index=2 * i + 1)
     plot_lda_cov(lda, splot)
     plt.axis('tight')
 
-    # QDA
-    qda = QDA()
+    # Quadratic Discriminant Analysis
+    qda = QuadraticDiscriminantAnalysis()
     y_pred = qda.fit(X, y, store_covariances=True).predict(X)
     splot = plot_data(qda, X, y, y_pred, fig_index=2 * i + 2)
     plot_qda_cov(qda, splot)
     plt.axis('tight')
-plt.suptitle('LDA vs QDA')
+plt.suptitle('Linear Discriminant Analysis vs Quadratic Discriminant Analysis')
 plt.show()
