+++ examples/decomposition/plot_pca_vs_lda.py
@@ -22,7 +22,7 @@ import matplotlib.pyplot as plt
 
 from sklearn import datasets
 from sklearn.decomposition import PCA
-from sklearn.lda import LDA
+from sklearn.discriminant_analysis import LinearDiscriminantAnalysis
 
 iris = datasets.load_iris()
 
@@ -33,7 +33,7 @@ target_names = iris.target_names
 pca = PCA(n_components=2)
 X_r = pca.fit(X).transform(X)
 
-lda = LDA(n_components=2)
+lda = LinearDiscriminantAnalysis(n_components=2)
 X_r2 = lda.fit(X, y).transform(X)
 
 # Percentage of variance explained for each components
