+++ sklearn/__init__.py
@@ -66,10 +66,9 @@ else:
                'linear_model', 'manifold', 'metrics', 'mixture', 'multiclass',
                'naive_bayes', 'neighbors', 'neural_network', 'pipeline',
                'preprocessing', 'qda', 'random_projection', 'semi_supervised',
-               'svm', 'tree',
+               'svm', 'tree', 'discriminant_analysis',
                # Non-modules:
                'clone']
-
 
 def setup_module(module):
     """Fixture for the tests to assure globally controllable seeding of RNGs"""
